Move posts to plugin dir and fix string to int in user extension
