+++ src/controller/articles.php
@@ -13,7 +13,7 @@ namespace ganstaz\gzo\src\controller;
 use ganstaz\gzo\src\controller\helper;
 use ganstaz\gzo\src\entity\manager as em;
 use ganstaz\gzo\src\form\form;
-use ganstaz\gzo\src\model\posts;
+use ganstaz\gzo\src\plugins\article\posts;
 use phpbb\event\dispatcher;
 use phpbb\exception\http_exception;
 use Symfony\Component\HttpFoundation\RedirectResponse;
